Replaced XNA by WPF as the GUI technology.
Motivation: XNA Content Projects are very difficult to set up in Visual Studio 2012, which heavily cripples XNA.
Note: WPF integration is difficult in Visual Studio 2012 Express for Web. Writing XAML in a designer seems pretty much out of the question. However, the client is just a small demo app, so this will not be a problem for the future.