Workaround for DragDrop.DoDragDrop blocking the UI thread.
The blocking happened when left mouse button was down, the mouse was still and a drop source moved to the mouse position. That would trigger a MouseMove event on the drop source which would call DragDrop.DoDragDrop which would block the UI thread that runs the animation (I suppose). The blocking can still happen if you click on the drag source and move so little that just one mouse event is fired. This is less likely to happen, however.